refactor(canvas): 拆分节点卡片样式与端口组件
